+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - #35 Fix error logging in the `Nest` channel
 - Add `publish/docker/latest` and `publish/docker/tag` make targets
+- #34 Add Buienradar weather description
 
 ### `0.8`